Publishes the first Rosalind entry.
It's worth noting that, despite being LHS, this won't actually work in
runhaskell or ghci right now (sigh).  This is due to [a bug][1] whereby the
Haskell interpreter sees our # symbols as preprocessor directives.

I'll revise this sometime tomorrow; in the spirit of actually blogging,
I'm happy with it going up in its current state.

[1]: http://hackage.haskell.org/trac/ghc/ticket/4836